% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derivatives_runner.R
\name{derivatives_runner_fun}
\alias{derivatives_runner_fun}
\title{Main Function to Run Derivatives}
\usage{
derivatives_runner_fun(storage_root, out_root, bbox_in, cpus)
}
\arguments{
\item{storage_root}{The root of the raw data}

\item{out_root}{The root of the location to write the results to}

\item{bbox_in}{The bounding box to calculate derivatives on.}

\item{cpus}{The number of cpus to use when calculating derivatives}
}
\description{
Main Function to Run Derivatives
}

