% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differencer.R
\name{differencer}
\alias{differencer}
\title{Difference periodized files between historical and past periods}
\usage{
differencer(out_path, future_path, historical_path, periods)
}
\arguments{
\item{future_path}{The path where the future periodized derivatives are}

\item{historical_path}{The path where the historical periodized derivatives are}

\item{periods}{The periods that were used in the periodizer}

\item{nc_file}{the netcdf file or opendap endpoint to use}

\item{data_path}{The path where the data files can be found}
}
\description{
Difference periodized files between historical and past periods
}

