% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensembler.R
\name{ensemble}
\alias{ensemble}
\title{Create ensemble of derivatives.}
\usage{
ensemble(nc_file, data_path, scenarios, gcm_scenarios, start, end)
}
\arguments{
\item{nc_file}{the netcdf file or opendap endpoint to use}

\item{data_path}{The path where the data files can be found}

\item{scenarios}{the scenarios that will go into the ensemble}

\item{gcm_scenarios}{the vecotr of gcm_scenario folders to consider}

\item{start}{A string start year}

\item{end}{A string end year}
}
\description{
Create ensemble of derivatives.
}

