% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/par_runner.R
\name{par_runner}
\alias{par_runner}
\title{Run dap_daily_stats for use with parSapply.}
\usage{
par_runner(start, end, bbox_in, thresholds, nc_file, NetCDF_output = TRUE, wd,
  data_path)
}
\arguments{
\item{start}{A string start year}

\item{end}{A string end year}

\item{bbox_in}{a vector of bbox indices per dap_daily_stats}

\item{thresholds}{a list of thresholds per dap_daily_stats}

\item{nc_file}{the netcdf file or opendap endpoint to use}

\item{NetCDF_output}{defaults to TRUE, geotiff if FALSE}

\item{wd}{The working directory to put derived results}

\item{data_path}{The path where the data files can be found}
}
\value{
fileNames the list of filenames that were created.
}
\description{
Run dap_daily_stats for use with parSapply.
}

