% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_dsg.R
\name{write_dsg}
\alias{write_dsg}
\title{Function to convert GDP output to NetCDF-DSG}
\usage{
write_dsg(gdp_output_root)
}
\arguments{
\item{gdp_output_root}{the path where gdp ouput gets dumped.}
}
\description{
Function to convert GDP output to NetCDF-DSG
}

