% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensembler.R
\name{ensemble}
\alias{ensemble}
\title{Create ensemble of derivatives.}
\usage{
ensemble(data_path, scenarios, start, end)
}
\arguments{
\item{data_path}{The path where the data files can be found}

\item{scenarios}{the scenarios that will go into the ensemble}

\item{start}{A string start year}

\item{end}{A string end year}
}
\description{
Create ensemble of derivatives.
}

